requirements.txt
discord.py
requests
python-dotenv
pandas